rag: basic rag example for tests
Create embeddings:
$ python3 load.py path/to/file.pdf

Run app:
$ python3 app.py

Ask the question about the document.

It is using chromadb to store embeddings on the disc. AI is considering
context and hisotry of previous qustion (only while app is running).

There is currently one session only, so it can be simplify to always
return ChatMessagesHistory() instance. Leav it like that for future
improvments.

Signed-off-by: Michal Swiatkowski <michal.swiatkowski@linux.intel.com>